added some update for the main project, version 1.50

+++ delivery_agency/src/main/java/service/impl/OrderProcessingPointServiceImpl.java
@@ -2,31 +2,36 @@ package service.impl;
 
 import dto.OrderProcessingPointDto;
 import entity.OrderProcessingPoint;
+import entity.Warehouse;
 import org.modelmapper.ModelMapper;
 import repository.OrderProcessingPointRepository;
 import repository.impl.OrderProcessingPointRepositoryImpl;
 import service.OrderProcessingPointService;
 
-import java.sql.SQLException;
 import java.util.List;
 
 public class OrderProcessingPointServiceImpl implements OrderProcessingPointService {
     private final OrderProcessingPointRepository orderProcessingPointRepository = new OrderProcessingPointRepositoryImpl();
+    private final ModelMapper modelMapper = new ModelMapper();
+
 
     @Override
-
     public OrderProcessingPoint save(OrderProcessingPointDto orderProcessingPointDto) {
         OrderProcessingPoint orderProcessingPoint = new OrderProcessingPoint();
         orderProcessingPoint.setId(orderProcessingPointDto.getId());
         orderProcessingPoint.setLocation(orderProcessingPointDto.getLocation());
-        orderProcessingPoint.setWarehouse(orderProcessingPoint.getWarehouse());
-        orderProcessingPoint.setExpectedOrders(orderProcessingPoint.getExpectedOrders());
-        orderProcessingPoint.setDispatchedOrders(orderProcessingPoint.getDispatchedOrders());
+        Warehouse warehouseToSave = new Warehouse();
+        warehouseToSave.setId(orderProcessingPointDto.getWarehouse().getId());
+        warehouseToSave.setLocation(orderProcessingPointDto.getWarehouse().getLocation());
+        orderProcessingPoint.setWarehouse(warehouseToSave);
         return orderProcessingPointRepository.save(orderProcessingPoint);
     }
 
     @Override
     public void delete(Long id) {
+        if (orderProcessingPointRepository.findOne(id) == null) {
+            throw new IllegalArgumentException("There is no processing Point with this Id!!! Cannot delete this client");
+        }
         orderProcessingPointRepository.delete(id);
     }
 
@@ -49,13 +54,13 @@ public class OrderProcessingPointServiceImpl implements OrderProcessingPointServ
     }
 
     @Override
-    public OrderProcessingPoint update(OrderProcessingPointDto orderProcessingPointDto) throws SQLException {
+    public OrderProcessingPoint update(OrderProcessingPointDto orderProcessingPointDto) {
         OrderProcessingPoint orderProcessingPoint = new OrderProcessingPoint();
         orderProcessingPoint.setId(orderProcessingPointDto.getId());
         orderProcessingPoint.setLocation(orderProcessingPointDto.getLocation());
-        orderProcessingPoint.setWarehouse(orderProcessingPoint.getWarehouse());
-        orderProcessingPoint.setExpectedOrders(orderProcessingPoint.getExpectedOrders());
-        orderProcessingPoint.setDispatchedOrders(orderProcessingPoint.getDispatchedOrders());
+        orderProcessingPoint.setWarehouse(modelMapper.map(orderProcessingPointDto.getWarehouse(), Warehouse.class));
+        //orderProcessingPoint.setExpectedOrders(orderProcessingPointDto.getExpectedOrders());
+        //orderProcessingPoint.setDispatchedOrders(orderProcessingPointDto.getDispatchedOrders());
         return orderProcessingPointRepository.update(orderProcessingPoint);
     }
